Extract save and load data logic into storage class
