+++ src/main/java/Main.java
@@ -1,7 +1,7 @@
 public class Main {
     public static void main(String[] args) {
         Xenon.greet();
-        Xenon chatbot = new Xenon();
+        Xenon chatbot = new Xenon("./data.txt");
         chatbot.chat();
         Xenon.exit();
     }
